migration fixed

+++ db/migrate/20180920213925_create_area_items.rb
@@ -3,7 +3,7 @@ class CreateAreaItems < ActiveRecord::Migration[5.2]
     create_table :area_items do |t|
       t.references :area, foreign_key: true
       t.references :item, foreign_key: true
-      t.integer :status
+      t.decimal(3,2) :status
 
       t.timestamps
     end
